Pacman and Ghosts won't check whether they can go in direction unless direction has changed

+++ MyPacman/MyPacman/World.cpp
@@ -44,7 +44,7 @@ void World::run(System &system, SystemGraphics &systemGraphics, SystemInput &sys
 	std::stringstream timeText;
 
 	int countedFrames = 0;
-	float avgFPS;
+	int avgFPS;
 	fpsTimer.start();
 	systemGraphics.createTextSprite("fpsTimer");
